Move `migrate_data` and `migrate_sessions` to migration crate
